🚚 Rename `index.ts` files to `main.ts`
