+++ src/vue.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'eslint/config';
 import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended';
-import mainRules from './rules/index.js';
+import mainRules from './rules/main.js';
 import vueRules from './rules/vue.js';
 
 const config = defineConfig([
